+++ NEWS.md
@@ -2,4 +2,5 @@
 
 ## Version 0.1.1
 
-Extend show for `ScaledVariable` and `SimpleExpression`.
+- Extend show for `ScaledVariable` and `LinearCombination`.
+- Rename `SimpleExpression` as `LinearCombination`.